Create PSET3RADIATIONEXPOSURE.PY
Assume that the curve function f is defined as follows:
def f(x):
    import math
    return 10*math.e**(math.log(0.5)/5.27 * x)
Test case 1:
>>> radiationExposure(0, 5, 1)
39.10318784326239
Test case 2:
>>> radiationExposure(5, 11, 1)
22.94241041057671
Test case 3:
>>> radiationExposure(0, 11, 1)
62.0455982538
Test case 4:
>>> radiationExposure(40, 100, 1.5)
0.434612356115
A note on these test cases: Your answers should be within 0.01 of the correct answer.
A Mathematical Note of Interest
The technique of finding the area under a curve is called integration. This comes to us from calculus. What we're doing in this problem is an approximation of finding the integral under the curve using the summation of rectangular areas known as a Riemann integral.
This approximation becomes more and more correct the smaller the width of the rectangles becomes.
So there you have it. If you've not learned calculus before, you've now got one of the basics - integration - covered!